refactored editor to removed old file
